add shared test support module
